StockPriceVisualisation/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
numpy
scikit-learn
yfinance
websockets
python-dotenv
matplotlib